week8/day37/orm_project/requirements.txt
bs4
requests
lxml
sqlalchemy
psycopg2-binary
